[수정] ModelController API URL 변경 및 모델 조회 테스트 코드 개선

+++ api-server/src/main/java/com/eavy/model/ModelRepository.java
@@ -2,6 +2,6 @@ package com.eavy.model;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ModelRepository extends JpaRepository<Model, Long> {
+public interface ModelRepository extends JpaRepository<Model, Integer> {
 
 }
